Refer: 폴더구조 및 기존 내용 리펙토링
