+++ Projects/App/Sources/SceneDelegate.swift
@@ -8,7 +8,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     
     func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
         // MARK: - Tuist 디자인 시스템 폰트 등록 - 
-        FontRegistrar.registerFont()
+        Pretendard.registerFont()
 
         guard let windowScene = (scene as? UIWindowScene) else { return }
         window = UIWindow(frame: windowScene.coordinateSpace.bounds)
