+++ Projects/Presentation/Sources/Auth/Login/LoginScene.swift
@@ -2,7 +2,7 @@ import UIKit
 
 public enum LoginScene {
     public static func create() -> UIViewController {
-        let viewModel = AuthViewModel()
+        let viewModel = LoginViewModel()
         let viewController = LoginViewController(viewModel: viewModel)
         return viewController
     }
