--- Projects/Presentation/Sources/View/ViewController.swift
@@ -1,27 +1,0 @@
-//
-//  ViewController.swift
-//  Login
-//
-//  Created by dgsw07 on 6/18/25.
-//
-
-import UIKit
-
-
-class ViewController: UIViewController {
-    
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        view.backgroundColor = .black
-        
-    }
-    
-    override func viewDidAppear(_ animated: Bool) {
-        super.viewDidAppear(true)
-        showAlert()
-    }
-    
-    func showAlert() {
-        
-    }
-}
