--- Projects/Presentation/Sources/ViewModel/SignUpViewModel.swift
@@ -1,1 +1,0 @@
-
